# Enable to set read_bytes for performance issue