@byroot 2022-08-18
All the connection code is gone in 5.0, it now rely on `redis-client`. Closing as stale.